Move chain stuff to abacus-core
