+++ rust/abacus-base/Cargo.toml
@@ -20,10 +20,6 @@ tracing-futures = "0.2"
 tracing-subscriber = { version = "0.3", features = ["json"] }
 rocksdb = "0.18"
 mockall = "0.10.2"
-num-traits = "0.2"
-num-derive = "0.3"
-strum = "0.24"
-strum_macros = "0.24"
 
 backtrace = { version = "0.3", optional = true }
 backtrace-oneline = { path = "../utils/backtrace-oneline", optional = true }
