+++ rust/abacus-core/Cargo.toml
@@ -17,7 +17,6 @@ sha3 = "0.9.1"
 lazy_static = "*"
 thiserror = "*"
 async-trait = { version = "0.1", default-features = false }
-num-traits = "0.2"
 maplit = "1.0"
 tokio = { version = "1", features = ["rt", "macros"] }
 tracing = "0.1"
@@ -28,6 +27,10 @@ eyre = "0.6"
 rocksdb = "0.18"
 bytes = { version = "1", features = ["serde"]}
 num = {version="0", features=["serde"]}
+num-traits = "0.2"
+num-derive = "0.3"
+strum = "0.24"
+strum_macros = "0.24"
 
 [dev-dependencies]
 abacus-base = { path = "../abacus-base" }
