+++ rust/abacus-core/src/chain.rs
@@ -1,7 +1,12 @@
 #![allow(missing_docs)]
 
+use std::str::FromStr;
+
 use eyre::Result;
+use num::FromPrimitive;
+use num_derive::FromPrimitive;
 use serde::{Deserialize, Serialize};
+use strum::EnumString;
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct Address(pub bytes::Bytes);
@@ -49,58 +54,204 @@ impl From<&'_ Address> for ethers::types::H160 {
     }
 }
 
-// /// Quick single-use macro to prevent typing domain and chain twice and risking
-// /// inconsistencies.
-// macro_rules! domain_and_chain {
-//     {$($domain:literal <=> $chain:literal,)*} => {
-//         /// Get the chain name from a domain id. Returns `None` if the `domain` is unknown.
-//         pub fn chain_from_domain(domain: u32) -> Option<&'static str> {
-//             match domain {
-//                 $( $domain => Some($chain), )*
-//                 _ => None
-//             }
-//         }
+/// All mainnet domains supported by Abacus.
+#[derive(FromPrimitive, EnumString, strum::Display, PartialEq, Eq, Debug)]
+#[strum(serialize_all = "lowercase")]
+pub enum AbacusMainnetDomain {
+    /// Ethereum domain ID, decimal ID 6648936
+    Ethereum = 0x657468,
 
-//         /// Get the domain id from a chain name. Expects `chain` to be a lowercase str.
-//         /// Returns `None` if the `chain` is unknown.
-//         pub fn domain_from_chain(chain: &str) -> Option<u32> {
-//             match chain {
-//                 $( $chain => Some($domain), )*
-//                 _ => None
-//             }
-//         }
-//     }
-// }
+    /// Polygon domain ID, decimal ID 1886350457
+    Polygon = 0x706f6c79,
 
-// // The unit test in this file `tests::json_mappings_match_code_map`
-// // tries to ensure some stability between the {chain} X {domain}
-// // mapping below with the agent configuration file.
-// domain_and_chain! {
-//     0x63656c6f <=> "celo",
-//     0x657468 <=> "ethereum",
-//     0x61766178 <=> "avalanche",
-//     0x706f6c79 <=> "polygon",
-//     1000 <=> "alfajores",
-//     43113 <=> "fuji",
-//     5 <=> "goerli",
-//     3000 <=> "kovan",
-//     80001 <=> "mumbai",
-//     6386274 <=> "arbitrum",
-//     6452067 <=> "bsc",
-//     28528 <=> "optimism",
-//     13371 <=> "test1",
-//     13372 <=> "test2",
-//     13373 <=> "test3",
-//     0x62732d74 <=> "bsctestnet",
-//     0x61722d72 <=> "arbitrumrinkeby",
-//     0x6f702d6b <=> "optimismkovan",
-//     0x61752d74 <=> "auroratestnet",
-//     0x6d6f2d61 <=> "moonbasealpha",
-// }
+    /// Avalanche domain ID, decimal ID 1635148152
+    Avalanche = 0x61766178,
+
+    /// Arbitrum domain ID, decimal ID 6386274
+    Arbitrum = 0x617262,
+
+    /// Optimism domain ID, decimal ID 28528
+    Optimism = 0x6f70,
+
+    /// BinanceSmartChain domain ID, decimal ID 6452067
+    #[strum(serialize = "bsc")]
+    BinanceSmartChain = 0x627363,
+
+    /// Celo domain ID, decimal ID 1667591279
+    Celo = 0x63656c6f,
+}
+
+impl From<AbacusMainnetDomain> for u32 {
+    fn from(domain: AbacusMainnetDomain) -> Self {
+        domain as u32
+    }
+}
+
+impl TryFrom<u32> for AbacusMainnetDomain {
+    type Error = eyre::Error;
+
+    fn try_from(domain_id: u32) -> Result<Self, Self::Error> {
+        FromPrimitive::from_u32(domain_id)
+            .ok_or_else(|| eyre::eyre!("Unknown mainnet domain ID {}", domain_id))
+    }
+}
+
+/// All testnet domains supported by Abacus.
+#[derive(FromPrimitive, EnumString, strum::Display, PartialEq, Eq, Debug)]
+#[strum(serialize_all = "lowercase")]
+pub enum AbacusTestnetDomain {
+    /// Ethereum testnet Goerli domain ID
+    Goerli = 5,
+    /// Ethereum testnet Kovan domain ID
+    Kovan = 3000,
+
+    /// Polygon testnet Mumbai domain ID
+    Mumbai = 80001,
+
+    /// Avalanche testnet Fuji domain ID
+    Fuji = 43113,
+
+    /// Arbitrum testnet ArbitrumRinkeby domain ID, decimal ID 1634872690
+    ArbitrumRinkeby = 0x61722d72,
+
+    /// Optimism testnet OptimismKovan domain ID, decimal ID 1869622635
+    OptimismKovan = 0x6f702d6b,
+
+    /// BSC testnet, decimal ID 1651715444
+    #[strum(serialize = "bsctestnet")]
+    BinanceSmartChainTestnet = 0x62732d74, // decimal 1651715444
+
+    /// Celo testnet Alfajores domain ID
+    Alfajores = 1000,
+
+    /// Moonbeam testnet MoonbaseAlpha domain ID, decimal ID 1836002657
+    MoonbaseAlpha = 0x6d6f2d61,
+}
+
+impl From<AbacusTestnetDomain> for u32 {
+    fn from(domain: AbacusTestnetDomain) -> Self {
+        domain as u32
+    }
+}
+
+impl TryFrom<u32> for AbacusTestnetDomain {
+    type Error = eyre::Error;
+
+    fn try_from(domain_id: u32) -> Result<Self, Self::Error> {
+        FromPrimitive::from_u32(domain_id)
+            .ok_or_else(|| eyre::eyre!("Unknown testnet domain ID {}", domain_id))
+    }
+}
+
+/// Local test chains (i.e. typically local hardhat nodes)
+#[derive(FromPrimitive, EnumString, strum::Display, PartialEq, Eq, Debug)]
+#[strum(serialize_all = "lowercase")]
+pub enum AbacusLocalTestChainDomain {
+    /// Test1 local chain
+    Test1 = 13371,
+    /// Test2 local chain
+    Test2 = 13372,
+    /// Test3 local chain
+    Test3 = 13373,
+}
+
+impl From<AbacusLocalTestChainDomain> for u32 {
+    fn from(domain: AbacusLocalTestChainDomain) -> Self {
+        domain as u32
+    }
+}
+
+impl TryFrom<u32> for AbacusLocalTestChainDomain {
+    type Error = eyre::Error;
+
+    fn try_from(domain_id: u32) -> Result<Self, Self::Error> {
+        FromPrimitive::from_u32(domain_id)
+            .ok_or_else(|| eyre::eyre!("Unknown local test chain domain ID {}", domain_id))
+    }
+}
+
+/// All domains supported by Abacus.
+#[derive(Debug, PartialEq, Eq)]
+pub enum AbacusDomain {
+    /// Mainnet domains.
+    Mainnet(AbacusMainnetDomain),
+    /// Testnet domains.
+    Testnet(AbacusTestnetDomain),
+    /// Local test domains (i.e. not public blockchains)
+    LocalTestChain(AbacusLocalTestChainDomain),
+}
+
+impl TryFrom<u32> for AbacusDomain {
+    type Error = eyre::Error;
+
+    fn try_from(domain_id: u32) -> Result<Self, Self::Error> {
+        if let Ok(mainnet_domain) = AbacusMainnetDomain::try_from(domain_id) {
+            Ok(Self::Mainnet(mainnet_domain))
+        } else if let Ok(testnet_domain) = AbacusTestnetDomain::try_from(domain_id) {
+            Ok(Self::Testnet(testnet_domain))
+        } else if let Ok(local_test_chain) = AbacusLocalTestChainDomain::try_from(domain_id) {
+            Ok(Self::LocalTestChain(local_test_chain))
+        } else {
+            Err(eyre::eyre!("Unknown domain ID {}", domain_id))
+        }
+    }
+}
+
+impl From<AbacusDomain> for u32 {
+    fn from(domain: AbacusDomain) -> Self {
+        match domain {
+            AbacusDomain::Mainnet(mainnet_domain) => mainnet_domain.into(),
+            AbacusDomain::Testnet(testnet_domain) => testnet_domain.into(),
+            AbacusDomain::LocalTestChain(local_test_chain) => local_test_chain.into(),
+        }
+    }
+}
+
+impl ToString for AbacusDomain {
+    fn to_string(&self) -> String {
+        match self {
+            AbacusDomain::Mainnet(mainnet_domain) => mainnet_domain.to_string(),
+            AbacusDomain::Testnet(testnet_domain) => testnet_domain.to_string(),
+            AbacusDomain::LocalTestChain(local_test_chain) => local_test_chain.to_string(),
+        }
+    }
+}
+
+impl FromStr for AbacusDomain {
+    type Err = strum::ParseError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        if let Ok(mainnet_domain) = AbacusMainnetDomain::from_str(s) {
+            Ok(Self::Mainnet(mainnet_domain))
+        } else if let Ok(testnet_domain) = AbacusTestnetDomain::from_str(s) {
+            Ok(Self::Testnet(testnet_domain))
+        } else if let Ok(local_test_chain) = AbacusLocalTestChainDomain::from_str(s) {
+            Ok(Self::LocalTestChain(local_test_chain))
+        } else {
+            Err(strum::ParseError::VariantNotFound)
+        }
+    }
+}
+
+/// Gets the name of the chain from a domain id.
+/// Returns None if the domain ID is not recognized.
+pub fn name_from_domain_id(domain_id: u32) -> Option<String> {
+    AbacusDomain::try_from(domain_id)
+        .ok()
+        .map(|domain| domain.to_string())
+}
+
+/// Gets the name of the chain from a domain id.
+/// Returns None if the domain ID is not recognized.
+pub fn domain_id_from_name(name: &'static str) -> Option<u32> {
+    AbacusDomain::from_str(name)
+        .ok()
+        .map(|domain| domain.into())
+}
 
 #[cfg(test)]
 mod tests {
-    use abacus_base::{Settings, chains::AbacusDomain};
+    use abacus_base::Settings;
     use config::{Config, File, FileFormat};
     use num_traits::identities::Zero;
     use std::collections::BTreeSet;
@@ -108,6 +259,11 @@ mod tests {
     use std::path::Path;
     use std::str::FromStr;
     use walkdir::WalkDir;
+
+    use crate::{
+        domain_id_from_name, name_from_domain_id, AbacusDomain, AbacusLocalTestChainDomain,
+        AbacusMainnetDomain, AbacusTestnetDomain,
+    };
 
     /// Relative path to the `abacus-monorepo/rust/config/`
     /// directory, which is where the agent's config files
@@ -251,10 +407,119 @@ mod tests {
 
         for ChainCoordinate { name, domain } in inbox_coords.iter().chain(outbox_coords.iter()) {
             assert_eq!(
-                AbacusDomain::try_from(domain.to_owned()).unwrap().to_string(),
+                AbacusDomain::try_from(domain.to_owned())
+                    .unwrap()
+                    .to_string(),
                 name.to_owned()
             );
-            assert_eq!(u32::from(AbacusDomain::from_str(name).unwrap()), domain.to_owned());
+            assert_eq!(
+                u32::from(AbacusDomain::from_str(name).unwrap()),
+                domain.to_owned()
+            );
         }
+    }
+
+    #[test]
+    fn mainnet_domain_strings() {
+        assert_eq!(
+            AbacusMainnetDomain::from_str("ethereum").unwrap(),
+            AbacusMainnetDomain::Ethereum,
+        );
+        assert_eq!(
+            AbacusMainnetDomain::Ethereum.to_string(),
+            "ethereum".to_string(),
+        );
+
+        // One where serialization has changed
+        assert_eq!(
+            AbacusMainnetDomain::from_str("bsc").unwrap(),
+            AbacusMainnetDomain::BinanceSmartChain,
+        );
+        assert_eq!(
+            AbacusMainnetDomain::BinanceSmartChain.to_string(),
+            "bsc".to_string(),
+        );
+
+        // Invalid name
+        assert!(AbacusMainnetDomain::from_str("foo").is_err());
+    }
+
+    #[test]
+    fn testnet_domain_strings() {
+        assert_eq!(
+            AbacusTestnetDomain::from_str("arbitrumrinkeby").unwrap(),
+            AbacusTestnetDomain::ArbitrumRinkeby,
+        );
+        assert_eq!(
+            AbacusTestnetDomain::ArbitrumRinkeby.to_string(),
+            "arbitrumrinkeby".to_string(),
+        );
+
+        // One where serialization has changed
+        assert_eq!(
+            AbacusTestnetDomain::from_str("bsctestnet").unwrap(),
+            AbacusTestnetDomain::BinanceSmartChainTestnet,
+        );
+        assert_eq!(
+            AbacusTestnetDomain::BinanceSmartChainTestnet.to_string(),
+            "bsctestnet".to_string(),
+        );
+
+        // Invalid name
+        assert!(AbacusTestnetDomain::from_str("foo").is_err());
+    }
+
+    #[test]
+    fn local_test_chain_domain_strings() {
+        assert_eq!(
+            AbacusLocalTestChainDomain::from_str("test1").unwrap(),
+            AbacusLocalTestChainDomain::Test1,
+        );
+        assert_eq!(
+            AbacusLocalTestChainDomain::Test1.to_string(),
+            "test1".to_string(),
+        );
+
+        // Invalid name
+        assert!(AbacusLocalTestChainDomain::from_str("foo").is_err());
+    }
+
+    #[test]
+    fn domain_strings() {
+        assert_eq!(
+            AbacusDomain::from_str("ethereum").unwrap(),
+            AbacusDomain::Mainnet(AbacusMainnetDomain::Ethereum),
+        );
+        assert_eq!(
+            AbacusDomain::Mainnet(AbacusMainnetDomain::Ethereum).to_string(),
+            "ethereum".to_string(),
+        );
+    }
+
+    #[test]
+    fn domain_ids() {
+        assert_eq!(
+            AbacusDomain::try_from(0x657468u32).unwrap(),
+            AbacusDomain::Mainnet(AbacusMainnetDomain::Ethereum),
+        );
+
+        assert_eq!(
+            u32::from(AbacusDomain::Mainnet(AbacusMainnetDomain::Ethereum)),
+            0x657468u32,
+        );
+    }
+
+    #[test]
+    fn test_name_from_domain_id() {
+        assert_eq!(name_from_domain_id(0x657468u32), Some("ethereum".into()),);
+
+        assert_eq!(name_from_domain_id(0xf00u32), None,);
+    }
+
+    #[test]
+    fn test_domain_id_from_name() {
+        assert_eq!(domain_id_from_name("ethereum"), Some(0x657468u32),);
+
+        assert_eq!(domain_id_from_name("foo"), None,);
     }
 }
